Test UDT fallback for parquet
